+++ tests/behat/behat-pantheon.yml
@@ -7,7 +7,6 @@ default:
       paths:
         - %paths.base%/features
       contexts:
-        - FeatureContext
         - Drupal\DrupalExtension\Context\DrushContext
         - Drupal\DrupalExtension\Context\DrupalContext
         - Drupal\DrupalExtension\Context\MinkContext
